cdk/requirements.txt
aws-cdk-lib==2.30.0
aws-cdk.aws-apigatewayv2-alpha==2.30.0a0
constructs>=10.0.0,<11.0.0
